+++ chunk-HW5TJ7QV.js
@@ -1,2 +1,0 @@
-import{Rb as n,Sb as i,mc as m,tb as o}from"./chunk-QC23KHUZ.js";import"./chunk-OKNOHCUX.js";var s=(()=>{class e{static \u0275fac=function(t){return new(t||e)};static \u0275cmp=o({type:e,selectors:[["app-utils-demo"]],decls:2,vars:0,template:function(t,a){t&1&&(n(0,"p"),m(1,"Coming Soon"),i())},encapsulation:2})}return e})();export{s as UtilsDemoComponent};
-//# sourceMappingURL=chunk-HW5TJ7QV.js.map